+++ src/gtt_activation_dialog.h
