+++ src/gtt_application_window.c
@@ -29,7 +29,7 @@
 
 #include <qof.h>
 
-#include "active-dialog.h"
+#include "gtt_activation_dialog.h"
 #include "app.h"
 #include "cur-proj.h"
 #include "gtt.h"
