+++ src/gtt_err_throw.c
@@ -18,10 +18,10 @@
 
 #include "config.h"
 
+#include "gtt_err_throw.h"
+
 #include <glib.h>
 #include <gnome.h>  /* needed only to define the _() macro */
-
-#include "err-throw.h"
 
 
 static GttErrCode err = GTT_NO_ERR;
