+++ src/gtt_err_throw.h
