+++ src/gtt_file_io.h
