+++ src/gtt_xml_read.c
@@ -25,7 +25,7 @@
 #include <stdlib.h>
 
 #include "cur-proj.h"
-#include "err-throw.h"
+#include "gtt_err_throw.h"
 #include "gtt.h"
 #include "proj.h"
 #include "proj_p.h"
